Revert "Client config not render blocking (#12300)"
This reverts commit 4ce7fc69878859e9ea3c65b6671b35f852326ee1.
