+++ packages/twenty-front/src/modules/settings/admin-panel/config-variables/hooks/useConfigVariableActions.ts
@@ -1,5 +1,6 @@
 import { useLingui } from '@lingui/react/macro';
 
+import { GET_CLIENT_CONFIG } from '@/client-config/graphql/queries/getClientConfig';
 import { GET_DATABASE_CONFIG_VARIABLE } from '@/settings/admin-panel/config-variables/graphql/queries/getDatabaseConfigVariable';
 import { SnackBarVariant } from '@/ui/feedback/snack-bar-manager/components/SnackBar';
 import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
@@ -47,6 +48,9 @@ export const useConfigVariableActions = (variableName: string) => {
               query: GET_DATABASE_CONFIG_VARIABLE,
               variables: { key: variableName },
             },
+            {
+              query: GET_CLIENT_CONFIG,
+            },
           ],
         });
       } else {
@@ -59,6 +63,9 @@ export const useConfigVariableActions = (variableName: string) => {
             {
               query: GET_DATABASE_CONFIG_VARIABLE,
               variables: { key: variableName },
+            },
+            {
+              query: GET_CLIENT_CONFIG,
             },
           ],
         });
@@ -88,6 +95,9 @@ export const useConfigVariableActions = (variableName: string) => {
           {
             query: GET_DATABASE_CONFIG_VARIABLE,
             variables: { key: variableName },
+          },
+          {
+            query: GET_CLIENT_CONFIG,
           },
         ],
       });
