+++ packages/twenty-front/src/testing/graphqlMocks.ts
@@ -124,9 +124,6 @@ export const graphqlMocks = {
         },
       });
     }),
-    http.get(`${REACT_APP_SERVER_BASE_URL}/client-config`, () => {
-      return HttpResponse.json(mockedClientConfig);
-    }),
     metadataGraphql.query(
       getOperationName(FIND_MANY_OBJECT_METADATA_ITEMS) ?? '',
       () => {
