+++ packages/twenty-server/src/engine/core-modules/captcha/interfaces/captcha.interface.ts
@@ -36,4 +36,5 @@ export type CaptchaModuleAsyncOptions = {
   ) => CaptchaModuleOptions | Promise<CaptchaModuleOptions> | undefined;
 } & Pick<ModuleMetadata, 'imports'> &
   Pick<FactoryProvider, 'inject'>;
+
 export type CaptchaValidateResult = { success: boolean; error?: string };
