--- packages/twenty-server/src/engine/core-modules/client-config/client-config.controller.ts
@@ -1,14 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-
-import { ClientConfig } from 'src/engine/core-modules/client-config/client-config.entity';
-import { ClientConfigService } from 'src/engine/core-modules/client-config/services/client-config.service';
-
-@Controller('/client-config')
-export class ClientConfigController {
-  constructor(private readonly clientConfigService: ClientConfigService) {}
-
-  @Get()
-  async getClientConfig(): Promise<ClientConfig> {
-    return this.clientConfigService.getClientConfig();
-  }
-}
