+++ packages/twenty-server/src/engine/core-modules/client-config/client-config.resolver.spec.ts
@@ -1,6 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { ClientConfigService } from 'src/engine/core-modules/client-config/services/client-config.service';
+import { DomainManagerService } from 'src/engine/core-modules/domain-manager/services/domain-manager.service';
+import { TwentyConfigService } from 'src/engine/core-modules/twenty-config/twenty-config.service';
 
 import { ClientConfigResolver } from './client-config.resolver';
 
@@ -12,10 +13,12 @@ describe('ClientConfigResolver', () => {
       providers: [
         ClientConfigResolver,
         {
-          provide: ClientConfigService,
-          useValue: {
-            getClientConfig: jest.fn(),
-          },
+          provide: TwentyConfigService,
+          useValue: {},
+        },
+        {
+          provide: DomainManagerService,
+          useValue: {},
         },
       ],
     }).compile();
